refactor: rename `render_ctx` to `render`
