+++ crates/frui_core/src/api/contexts/mod.rs
@@ -1,7 +1,7 @@
 use crate::app::tree::WidgetNodeRef;
 
 pub mod build_ctx;
-pub mod render_ctx;
+pub mod render;
 
 #[repr(transparent)]
 pub struct Context {
