+++ crates/frui_core/src/app/tree/mod.rs
@@ -11,7 +11,7 @@ use druid_shell::IdleToken;
 
 use crate::{
     api::{
-        contexts::{render_ctx::LayoutCtxOS, Context},
+        contexts::{render::LayoutCtxOS, Context},
         local_key::LocalKeyAny,
         pointer_events::events::PointerEvent,
         IntoWidgetPtr, WidgetPtr,
